refactor feedbackScreen stories
